Fixed recorder data containing Tensors
Somewhere along the way, self.recorder ended up with an element that was a Tensor, not a float, causing numpy.save() to crash. Trying to find where the Tensor is stored may be quite difficult to do.

Instead, I added a few nifty helper-functions to detect whether or not self.recorder contains any Tensors and force them to be floats before calling numpy.save().