Refactor: black formatting
I used the Black code formatter to help alleviate a lot of the very long
lines that make it difficult to parse the code in some of the files.
Now, these files are more readable and maintainable - none of the
functionality has been changed. Note: I've only gone through the agent
files for now, since we are still fixing some of the other files
involving Isaac Gym.