Fixed error involving a single critic minimum
In order to run AgentDDPG (a single critic DRL algorithm), the code
needs to be able to handle whether the algorithm makes use of a single
critic or multiple critics. Originally, it would throw an error when
running AgentDDPG, since it only uses a single critic, but this has
been patched.