Template task for creating new RL environment
Running Isaac Gym on a custom RL environment (for example, an OpenAI Gym
environment or a financial-trading environment) requires setting up a
class for that environment, which inherits from `VecTask`. Because of
this, the training architecture that ElegantRL uses will likely have to
accomodate this detail. Using `train_and_evaluate()` or
`train_and_evaluate_mp()` may be insufficient, so I may have to create a
new method `train_and_evaluate_isaac()` (?) which accomodates Isaac Gym
as such.