Testing simulation setup in Isaac Gym
I'm going through the Isaac Gym documentation, trying to create a
minimalist environment to train on. All the example environments they
have provided use all sorts of fancy assets in URDF and MJCF, but RL
tasks do not have to be that complicated: just a simple environment that
can be stepped through, reset, etc.