Added support for get_isaac_env_args()
One thing that makes it a bit tricky to use Isaac Gym is knowing the
state and action dimensions of all the various environments (12 of
them, I think), so I've implemented a function get_isaac_env_args() that
returns all the necessary parameters as long as you just specify the
name. With this, the end user just has to figure out the environment
name they'd like to train on as well as the training function they'd
like to use and they're off to the races!