Fix dead link to `elegantrl_helloworld`
The link `ElegantRL-Helloworld` in the intro at the latest documentation (https://elegantrl.readthedocs.io/en/latest/helloworld/intro.html) is broken. I believe it was renamed in the repo but the change wasn't reflected in the docs. This fixes the broken link to point to the new remote url `https://github.com/AI4Finance-Foundation/ElegantRL/tree/master/helloworld`.